+++ prisma/schema.prisma
@@ -28,7 +28,7 @@ model Investment {
   company      Company  @relation(fields: [companyId], references: [id], onDelete: Cascade)
   companyId    String   @db.ObjectId
   rawName      String
-  coupon       String
+  coupon       Float
   maturityDate DateTime
   quantity     Float
   costVal      Float
