+++ src/components/index.ts
@@ -1,13 +1,25 @@
+// Buttons
 import AuthButton from './Buttons/AuthButton'
-import ListItem from './Displays/ListItem'
-import NumberCircle from './Text/NumberCircle'
-import HighlightedTitle from './Text/HighlightedTitle'
 import NavButton from './Buttons/NavButton'
-import NavBar from './Nav/NavBar'
 import PrimaryButton from './Buttons/PrimaryButton'
 import ToTopButton from './Buttons/ToTopButton'
+// Charts
+import LandingDonutChart from './Charts/LandingDonutChart'
+// Displays
+import ListItem from './Displays/ListItem'
+import LinkDisplay from './Displays/LinkDisplay'
+import DropDownCard from './Displays/DropDownCard'
 import Carousel from './Displays/Carousel/Carousel'
+// Nav
+import NavBar from './Nav/NavBar'
+import SecondaryNavBar from './Nav/SecondaryNavBar'
+// Selects
 import Select from './Selects/Select'
+// Text
+import NumberCircle from './Text/NumberCircle'
+import HighlightedTitle from './Text/HighlightedTitle'
+import CapitalizedTitle from './Text/CapitalizedTitle'
+import ShadowTitle from './Text/ShadowTitle'
 
 export {
   AuthButton,
@@ -20,4 +32,10 @@ export {
   ToTopButton,
   Carousel,
   Select,
+  LandingDonutChart,
+  CapitalizedTitle,
+  LinkDisplay,
+  DropDownCard,
+  SecondaryNavBar,
+  ShadowTitle,
 }
