+++ src/info/home.json
@@ -1,5 +1,4 @@
 {
-  "example": "an awesome example",
   "ourRequests": [
     {
       "header": "We request Chancellor Jones, President Tim Killeen and University of Illinois Foundation CFO Jim Moore publish explicit language detailing fossil fuel divestment by the University of Illinois System.",
