+++ src/sections/home/SampleSection.tsx
@@ -1,6 +1,6 @@
 import type { FC } from 'react'
 
-import { HighlightedTitle } from '../components'
+import { HighlightedTitle } from '../../components'
 
 interface InputProps {
   text: string
