+++ src/sections/index.tsx
@@ -1,0 +1,15 @@
+import InstitutionalDivestments from './home/InstitutionalDivestments'
+import Landing from './home/Landing'
+import OurRequest from './home/OurRequests'
+import RefuteUISResponse from './home/RefuteUISResponse'
+import SampleSection from './home/SampleSection'
+import TimelineSection from './home/Timeline'
+
+export {
+  InstitutionalDivestments,
+  Landing,
+  OurRequest,
+  RefuteUISResponse,
+  SampleSection,
+  TimelineSection,
+}
